UPDATE the truncation parameters as long objects

+++ src/main/java/com/telefonica/iot/cygnus/sinks/OrionMongoBaseSink.java
@@ -41,7 +41,7 @@ public abstract class OrionMongoBaseSink extends OrionSink {
     protected String collectionPrefix;
     protected boolean shouldHash;
     protected MongoBackendImpl backend;
-    protected int dataExpiration;
+    protected long dataExpiration;
 
     /**
      * Gets the mongo hosts. It is protected since it is used by the tests.
@@ -114,7 +114,7 @@ public abstract class OrionMongoBaseSink extends OrionSink {
         LOGGER.debug("[" + this.getName() + "] Reading configuration (collection_prefix=" + collectionPrefix + ")");
         shouldHash = context.getBoolean("should_hash", false);
         LOGGER.debug("[" + this.getName() + "] Reading configuration (should_hash=" + shouldHash + ")");
-        dataExpiration = context.getInteger("data_expiration", 0);
+        dataExpiration = context.getLong("data_expiration", 0L);
         LOGGER.debug("[" + this.getName() + "] Reading configuration (data_expiraton=" + dataExpiration + ")");
         super.configure(context);
     } // configure
